refactor(IMG) : IMG 패키지 이름 변경 및 위치 이동
