+++ src/main/java/com/multi/travel/domain/place/img/dto/PlaceImageResDTO.java
@@ -1,4 +1,4 @@
-package com.multi.travel.domain.place_img.dto;
+package com.multi.travel.domain.place.img.dto;
 
 /*
  * Please explain the class!!!
@@ -9,7 +9,7 @@ package com.multi.travel.domain.place_img.dto;
  */
 
 
-import com.multi.travel.domain.place_img.entity.PlaceImage;
+import com.multi.travel.domain.place.img.entity.PlaceImage;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
