+++ src/main/java/com/multi/travel/domain/place/service/PlaceService.java
@@ -14,8 +14,8 @@ import com.multi.travel.domain.place.dto.*;
 import com.multi.travel.domain.place.entity.Place;
 import com.multi.travel.domain.place.enums.PlaceStatus;
 import com.multi.travel.domain.place.repository.PlaceRepository;
-import com.multi.travel.domain.place_img.entity.PlaceImage;
-import com.multi.travel.domain.place_img.repository.PlaceImageRepository;
+import com.multi.travel.domain.place.img.entity.PlaceImage;
+import com.multi.travel.domain.place.img.repository.PlaceImageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
